% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-help.R
\name{ax.celsiuslab}
\alias{ax.celsiuslab}
\title{Temperature axis label}
\usage{
ax.celsiuslab(label = "Temperature", font = par()$font.axis)
}
\arguments{
\item{label}{the label of the temperature variable, defaults to
"Temperature".}

\item{font}{An integer which specifies which font to use for text. Defaults
to the axis font currently set by \code{par}; see \code{?par} for the
available options.}
}
\value{
A language object with the axis label.
}
\description{
Provides the axis label for temperature variables in units of degree
Celsius.
}
\examples{
plot(1 : 10, xlab = ax.celsiuslab(),
             ylab = ax.celsiuslab("Sea surface temperature", font = 2))
}
\seealso{
\code{\link{bquote}}
}
\author{
Thomas Münch
}
