% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-help.R
\name{ax.freqlab}
\alias{ax.freqlab}
\title{Frequency axis label}
\usage{
ax.freqlab(label = if (!inverse) "Frequency" else "Time period",
  unit = "yr", font = par()$font.axis, inverse = FALSE)
}
\arguments{
\item{label}{the label for the frequency axis; defaults to "Frequency" for
\code{inverse = FALSE}, else to "Time period".}

\item{unit}{the time unit of the corresponding frequency axis; defaults to
years ("yr").}

\item{font}{An integer which specifies which font to use for text. Defaults
to the axis font currently set by \code{par}; see \code{?par} for the
available options.}

\item{inverse}{logical; if \code{TRUE}, use inverse frequency units,
i.e. time periods.}
}
\value{
A language object with the axis label.
}
\description{
Provides the frequency axis label for spectral plots in optional units.
}
\examples{
plot(1 : 10, xlab = ax.freqlab(),
             ylab = ax.freqlab(unit = "s", inverse = TRUE, font = 2))
}
\author{
Thomas Münch
}
