% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-help.R
\name{ax.isolab}
\alias{ax.isolab}
\title{Isotope delta notation label}
\usage{
ax.isolab(x = "oxy", font = par()$font.axis)
}
\arguments{
\item{x}{either a character string to set the isotopic species, or a general
character string to use as axis label; see details. Defaults to oxygen
isotopic composition.}

\item{font}{An integer which specifies which font to use for text. Defaults
to the axis font currently set by \code{par}; see \code{?par} for the
available options.}
}
\value{
A language object with the axis label.
}
\description{
Provides the axis label for isotopic composition in delta notation and units
of permil.
}
\details{
If \code{x} is set to "oxy" or "dtr", axis labels are formatted according to
the delta notation for oxygen or deuterium isotopic composition,
respectively; else, \code{x} is used directly as the axis label variable.
}
\examples{
plot(1 : 10, xlab = ax.isolab(), ylab = ax.isolab("dtr"))
plot(1 : 10, xlab = ax.isolab(font = 2),
             ylab = ax.isolab("d-excess", font = 2))
}
\seealso{
\code{\link{bquote}}
}
\author{
Thomas Münch
}
