% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-help.R
\name{ax.psdlab}
\alias{ax.psdlab}
\title{Power spectral density axis label}
\usage{
ax.psdlab(label = "Power spectral density", v.unit = "‰",
  t.unit = "yr", font = par()$font.axis)
}
\arguments{
\item{label}{the label for the spectrum axis; defaults to "Power spectral
density".}

\item{v.unit}{the unit of the spectral variable; defaults to permil.}

\item{t.unit}{the frequency unit of the spectrum; defaults to years ("yr").}

\item{font}{An integer which specifies which font to use for text. Defaults
to the axis font currently set by \code{par}; see \code{?par} for the
available options.}
}
\value{
A language object with the axis label.
}
\description{
Provides the power spectral density axis label for spectral plots in
optional units.
}
\examples{
plot(1 : 10, xlab = ax.freqlab(), ylab = ax.psdlab())
}
\author{
Thomas Münch
}
