% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-help.R
\name{bquote.font}
\alias{bquote.font}
\title{Font selection for bquote}
\usage{
bquote.font(expr, font = 1)
}
\arguments{
\item{expr}{A language object.}

\item{font}{An integer which specifies which font to use for text; see
\code{?par} for the available options.}
}
\value{
A language object.
}
\description{
A wrapper function for \code{\link{bquote}} to select text font according to
the function argument \code{font}.
}
\author{
Thomas Münch
}
