% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-help.R
\name{ColorPal}
\alias{ColorPal}
\title{Default color palette}
\usage{
ColorPal(RCB.pal = "RdYlBu", n = 10)
}
\arguments{
\item{RCB.pal}{a palette name from the RColorBrewer lists; defaults to the
diverging \code{RdYlBu} palette.}

\item{n}{number of different colours in the palette, minimum 3, maximum
depending on the palette. Default is \code{10} for \code{RdYlBu}.}
}
\value{
a function that takes an integer argument (the required number of
colours) and returns a character vector of colours.
}
\description{
Wrapper to obtain my default colour palette for filled contour plots, etc.
}
\seealso{
\code{\link{colorRampPalette}};
\code{\link{RColorBrewer::brewer.pal}}
}
\author{
Thomas Münch
}
