% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.legend.R
\name{my.legend}
\alias{my.legend}
\title{Add Legends to Plots}
\usage{
my.legend(x, y = NULL, legend, fill = NULL, col = par("col"),
  border = "black", lty, lwd, pch, angle = 45, density = NULL,
  bty = "o", bg = par("bg"), box.lwd = par("lwd"),
  box.lty = par("lty"), box.col = par("fg"), pt.bg = NA, cex = 1,
  pt.cex = cex, pt.lwd = lwd, xjust = 0, yjust = 1,
  x.intersp = 1, y.intersp = 1, adj = c(0, 0.5), text.width = NULL,
  text.col = par("col"), text.font = NULL, merge = do.lines &&
  has.pch, trace = FALSE, plot = TRUE, ncol = 1, horiz = FALSE,
  title = NULL, inset = 0, xpd, title.col = text.col,
  title.adj = 0.5, seg.len = 2, end.pch = FALSE, pch.xoff = 0.2)
}
\arguments{
\item{x}{the x and y co-ordinates to be used to position the legend.
    They can be specified by keyword or in any way which is accepted by
    \code{\link{xy.coords}}: See \sQuote{Details}.}

\item{y}{the x and y co-ordinates to be used to position the legend.
    They can be specified by keyword or in any way which is accepted by
    \code{\link{xy.coords}}: See \sQuote{Details}.}

\item{legend}{a character or \link{expression} vector
    of length \eqn{\ge 1} to appear in the legend.  Other
    objects will be coerced by \code{\link{as.graphicsAnnot}}.}

\item{fill}{if specified, this argument will cause boxes filled
    with the specified colors (or shaded in the specified colors)
    to appear beside the legend text.}

\item{col}{the color of points or lines appearing in the legend.}

\item{border}{the border color for the boxes (used only if
    \code{fill} is specified).}

\item{lty}{the line types and widths for lines appearing in the
    legend.  One of these two \emph{must} be specified for line drawing.}

\item{lwd}{the line types and widths for lines appearing in the
    legend.  One of these two \emph{must} be specified for line drawing.}

\item{pch}{the plotting symbols appearing in the legend, as
    numeric vector or a vector of 1-character strings (see
    \code{\link{points}}). Unlike \code{points}, this can all be
    specified as a single multi-character string.  \emph{Must} be
    specified for symbol drawing.}

\item{angle}{angle of shading lines.}

\item{density}{the density of shading lines, if numeric and
    positive. If \code{NULL} or negative or \code{NA} color filling
    is assumed.}

\item{bty}{the type of box to be drawn around the legend.  The allowed
    values are \code{"o"} (the default) and \code{"n"}.}

\item{bg}{the background color for the legend box.  (Note that this is
    only used if \code{bty != "n"}.)}

\item{box.lwd}{the line type, width and color for
    the legend box (if \code{bty = "o"}).}

\item{box.lty}{the line type, width and color for
    the legend box (if \code{bty = "o"}).}

\item{box.col}{the line type, width and color for
    the legend box (if \code{bty = "o"}).}

\item{pt.bg}{the background color for the \code{\link{points}},
    corresponding to its argument \code{bg}.}

\item{cex}{character expansion factor \bold{relative} to current
    \code{par("cex")}.  Used for text, and provides the default for
    \code{pt.cex}.}

\item{pt.cex}{expansion factor(s) for the points.}

\item{pt.lwd}{line width for the points, defaults to the one for
    lines, or if that is not set, to \code{par("lwd")}.}

\item{xjust}{how the legend is to be justified relative to the legend
    x location.  A value of 0 means left justified, 0.5 means centered
    and 1 means right justified.}

\item{yjust}{the same as \code{xjust} for the legend y location.}

\item{x.intersp}{character interspacing factor for horizontal (x) spacing.}

\item{y.intersp}{the same for vertical (y) line distances.}

\item{adj}{numeric of length 1 or 2; the string adjustment for legend
    text.  Useful for y-adjustment when \code{labels} are
    \link{plotmath} expressions.}

\item{text.width}{the width of the legend text in x (\code{"user"})
    coordinates.  (Should be a single positive number even for a reversed
    x axis.)  Defaults to the proper value computed by
    \code{\link{strwidth}(legend)}.}

\item{text.col}{the color used for the legend text.}

\item{text.font}{the font used for the legend text, see \code{\link{text}}.}

\item{merge}{logical; if \code{TRUE}, merge points and lines but
    not filled boxes.  Defaults to \code{TRUE} if there are points and lines.}

\item{trace}{logical; if \code{TRUE}, shows how \code{legend} does all
    its magical computations.}

\item{plot}{logical.  If \code{FALSE}, nothing is plotted but the
    sizes are returned.}

\item{ncol}{the number of columns in which to set the legend items
    (default is 1, a vertical legend).}

\item{horiz}{logical; if \code{TRUE}, set the legend horizontally
    rather than vertically (specifying \code{horiz} overrides the \code{ncol}
    specification).}

\item{title}{a character string or length-one expression giving a
    title to be placed at the top of the legend.  Other objects will be
    coerced by \code{\link{as.graphicsAnnot}}.}

\item{inset}{inset distance(s) from the margins as a fraction of the
    plot region when legend is placed by keyword.}

\item{xpd}{if supplied, a value of the \link{graphical parameter} \code{xpd}
    to be used while the legend is being drawn.}

\item{title.col}{color for \code{title}.}

\item{title.adj}{horizontal adjustment for \code{title}: see the help
    for \code{\link{par}("adj")}.}

\item{seg.len}{the length of lines drawn to illustrate \code{lty}
    and/or \code{lwd} (in units of character widths).}

\item{end.pch}{if \code{TRUE}, add the point characters specified by
\code{pch} to both end points of the legend lines specified by
\code{lty}. Defaults to \code{FALSE} which reproduces the base behaviour.}

\item{pch.xoff}{horizontal adjustment of the point characters at the end
points of the legend lines to improve visual appearance; in units of
character widths (defaults to 0.2).}
}
\value{
the value returned by \code{legend}.
}
\description{
This is a modified version of the base R function \code{legend()} to add
legends to a point plot connected by lines. For such a plot, the base
version adds the point character in the middle of the legend line; in this
modified version, the point characters are added to both ends of the legend
line.
}
\seealso{
\code{\link{legend}}
}
\author{
Thomas Münch
}
