% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2png.R
\name{pdf2png}
\alias{pdf2png}
\title{pdf to png conversion}
\source{
https://robservatory.com/use-sips-to-quickly-easily-and-freely-convert-image-files/
}
\usage{
pdf2png(file)
}
\arguments{
\item{file}{the root file name of the pdf file to convert, i.e. the filename
without the \code{.pdf} extension.}
}
\value{
the default output from \code{\link{system}}.
}
\description{
Convert an existing pdf file to a png file using the Mac OS X command line
tool \code{sips}.
}
\author{
Thomas Münch
}
