% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-help.R
\name{PlotStamp}
\alias{PlotStamp}
\title{Add plot session information}
\usage{
PlotStamp(script = "", add.info = NULL, incr = 0.01)
}
\arguments{
\item{script}{the name of the plot creating script.}

\item{add.info}{additional information to be added (optional).}

\item{incr}{fraction of the user coordinate range to use as incremental
distance from the default position of the information text in the upper
right corner of the plot region.}
}
\description{
Add session information to an existing plot: script name, date of plot
creation, etc.
}
\examples{
plot(1 : 10, type = "l")
PlotStamp(script = "interactive",
          add.info = "Illustrated usage of PlotStamp function")
}
\author{
Thomas Münch
}
