% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quartz.R
\name{Quartz}
\alias{Quartz}
\title{Open quartz device}
\usage{
Quartz(file = NULL, type = "native", height = 6, width = 8, ...)
}
\arguments{
\item{file}{path to a file for storing a hardcopy of the plot including a
supported file extension to set the \code{type} of output (e.g. ".pdf" or
".png"); i.e. the function extracts the extension from \code{file} and uses
it as the \code{type} argument for the call to
\code{\link{quartz}}. Defaults to \code{NULL} for on-screen plotting.}

\item{type}{the type of output to use. Defaults to \code{"native"} for
on-screen plotting. If \code{file} is not \code{NULL}, \code{type} is
determined from its extension.}

\item{height}{the height of the plotting area in inches.  Default ‘6’.}

\item{width}{the width of the plotting area in inches.  Default ‘8’.}

\item{...}{further arguments passed on to \code{\link{quartz}}.}
}
\description{
Wrapper to open a quartz device with my default dimensions and graphical
parameters for on-screen plotting or saving to a file.
}
\examples{

# Create an empty on-screen quartz device
Quartz()

# Store empty plot in pdf format in local directory
\dontrun{
Quartz(file = file.path(getwd(), "test-quartz.pdf"))
dev.off()
}
}
\seealso{
\code{\link{quartz}}
}
\author{
Thomas Münch
}
