% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-help.R
\name{rel.hist}
\alias{rel.hist}
\title{Relative histograms}
\usage{
rel.hist(x, breaks = "Sturges", max.normalize = FALSE, add = FALSE,
  ...)
}
\arguments{
\item{x}{a vector of values for which the histogram is desired.}

\item{breaks}{one of:
    \itemize{
      \item a vector giving the breakpoints between histogram cells,
      \item a function to compute the vector of breakpoints,
      \item a single number giving the number of cells for the histogram,
      \item a character string naming an algorithm to compute the
      number of cells (see \sQuote{Details}),
      \item a function to compute the number of cells.
    }
    In the last three cases the number is a suggestion only; as the
    breakpoints will be set to \code{\link{pretty}} values, the number
    is limited to \code{1e6} (with a warning if it was larger).  If
    \code{breaks} is a function, the \code{x} vector is supplied to it
    as the only argument (and the number of breaks is only limited by
    the amount of available memory).
  }

\item{max.normalize}{if \code{TRUE}, normalize the counts to their maximum
value; else normalize with respect to the sum of the counts (the default).}

\item{add}{logical. If ‘TRUE’, only the bars are added to the current plot.}

\item{...}{further arguments passed on to \code{plot}.}
}
\value{
Invisibly, the resulting object of class \code{"histogram"}.
}
\description{
Compute and plot relative histograms with respect to the total or the
maximum of the counts.
}
\examples{
x <- rnorm(1000)

# **counts** sum up to one:
h <- rel.hist(x)
sum(h$counts)
# per default, histogram **area** sums up to one:
h <- hist(x, plot = FALSE)
plot(h, freq = FALSE)
sum(h$density) * diff(h$breaks)[1]

# counts normalized to maximum value:
rel.hist(x, max.normalize = TRUE)
}
\seealso{
\code{\link{hist}}
}
\author{
Thomas Münch
}
