% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceDir.R
\name{sourceDir}
\alias{sourceDir}
\title{Source scripts from directory.}
\source{
see the respective example in \code{?source}.
}
\usage{
sourceDir(path, trace = TRUE, local = FALSE, ...)
}
\arguments{
\item{path}{the path to the directory to source.}

\item{trace}{if \code{TRUE}, the names of the sourced files are printed to
the terminal in order to facilitate tracing of errors upon sourcing.}

\item{local}{\code{TRUE}, \code{FALSE} or an environment, determining
    where the parsed expressions are evaluated.  \code{FALSE} (the
    default) corresponds to the user's workspace (the global
    environment) and \code{TRUE} to the environment from which
    \code{source} is called.}

\item{...}{further arguments passed on to \code{source}.}
}
\description{
Source all \code{R}-related scripts from  a given directory.
}
\details{
The function looks for all files with extension matching \code{R}, \code{r},
\code{S}, \code{s}, \code{Q} and \code{q} in the given directory and sources
them with \code{source()}.
}
\seealso{
\code{\link{source}}
}
\author{
Thomas Münch
}
